refactor: consolidate public chat snapshot mutations with cache invalidation
